+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/DOMWheelEvent.java
@@ -29,11 +29,11 @@ public boolean altKey() {
 }
 
 public int clientX() {
-	return OS.objc_msgSend(this.id, OS.sel_clientX);
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_clientX);
 }
 
 public int clientY() {
-	return OS.objc_msgSend(this.id, OS.sel_clientY);
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_clientY);
 }
 
 public boolean ctrlKey() {
@@ -49,7 +49,7 @@ public boolean shiftKey() {
 }
 
 public int wheelDelta() {
-	return OS.objc_msgSend(this.id, OS.sel_wheelDelta);
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_wheelDelta);
 }
 
 }
