+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSAttributedString.java
@@ -24,12 +24,12 @@ public NSAttributedString(id id) {
 	super(id);
 }
 
-public NSDictionary attributesAtIndex(int location, int range, NSRange rangeLimit) {
-	int result = OS.objc_msgSend(this.id, OS.sel_attributesAtIndex_longestEffectiveRange_inRange_, location, range, rangeLimit);
+public NSDictionary attributesAtIndex(int /*long*/ location, int /*long*/ range, NSRange rangeLimit) {
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_attributesAtIndex_longestEffectiveRange_inRange_, location, range, rangeLimit);
 	return result != 0 ? new NSDictionary(result) : null;
 }
 
-public NSRange doubleClickAtIndex(int location) {
+public NSRange doubleClickAtIndex(int /*long*/ location) {
 	NSRange result = new NSRange();
 	OS.objc_msgSend_stret(result, this.id, OS.sel_doubleClickAtIndex_, location);
 	return result;
@@ -44,12 +44,12 @@ public void drawInRect(NSRect rect) {
 }
 
 public NSAttributedString initWithString(NSString str, NSDictionary attrs) {
-	int result = OS.objc_msgSend(this.id, OS.sel_initWithString_attributes_, str != null ? str.id : 0, attrs != null ? attrs.id : 0);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_initWithString_attributes_, str != null ? str.id : 0, attrs != null ? attrs.id : 0);
 	return result == this.id ? this : (result != 0 ? new NSAttributedString(result) : null);
 }
 
-public int nextWordFromIndex(int location, boolean isForward) {
-	return OS.objc_msgSend(this.id, OS.sel_nextWordFromIndex_forward_, location, isForward);
+public int /*long*/ nextWordFromIndex(int /*long*/ location, boolean isForward) {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_nextWordFromIndex_forward_, location, isForward);
 }
 
 public NSSize size() {
@@ -59,21 +59,21 @@ public NSSize size() {
 }
 
 public NSAttributedString attributedSubstringFromRange(NSRange range) {
-	int result = OS.objc_msgSend(this.id, OS.sel_attributedSubstringFromRange_, range);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_attributedSubstringFromRange_, range);
 	return result == this.id ? this : (result != 0 ? new NSAttributedString(result) : null);
 }
 
 public NSAttributedString initWithString(NSString str) {
-	int result = OS.objc_msgSend(this.id, OS.sel_initWithString_, str != null ? str.id : 0);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_initWithString_, str != null ? str.id : 0);
 	return result == this.id ? this : (result != 0 ? new NSAttributedString(result) : null);
 }
 
-public int length() {
-	return OS.objc_msgSend(this.id, OS.sel_length);
+public int /*long*/ length() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_length);
 }
 
 public NSString string() {
-	int result = OS.objc_msgSend(this.id, OS.sel_string);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_string);
 	return result != 0 ? new NSString(result) : null;
 }
