+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSBezierPath.java
@@ -32,15 +32,15 @@ public void appendBezierPath(NSBezierPath path) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPath_, path != null ? path.id : 0);
 }
 
-public void appendBezierPathWithArcWithCenter(NSPoint center, float radius, float startAngle, float endAngle) {
+public void appendBezierPathWithArcWithCenter(NSPoint center, float /*double*/ radius, float /*double*/ startAngle, float /*double*/ endAngle) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithArcWithCenter_radius_startAngle_endAngle_, center, radius, startAngle, endAngle);
 }
 
-public void appendBezierPathWithArcWithCenter(NSPoint center, float radius, float startAngle, float endAngle, boolean clockwise) {
+public void appendBezierPathWithArcWithCenter(NSPoint center, float /*double*/ radius, float /*double*/ startAngle, float /*double*/ endAngle, boolean clockwise) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithArcWithCenter_radius_startAngle_endAngle_clockwise_, center, radius, startAngle, endAngle, clockwise);
 }
 
-public void appendBezierPathWithGlyphs(int glyphs, int count, NSFont font) {
+public void appendBezierPathWithGlyphs(int /*long*/ glyphs, int /*long*/ count, NSFont font) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithGlyphs_count_inFont_, glyphs, count, font != null ? font.id : 0);
 }
 
@@ -52,22 +52,22 @@ public void appendBezierPathWithRect(NSRect rect) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithRect_, rect);
 }
 
-public void appendBezierPathWithRoundedRect(NSRect rect, float xRadius, float yRadius) {
+public void appendBezierPathWithRoundedRect(NSRect rect, float /*double*/ xRadius, float /*double*/ yRadius) {
 	OS.objc_msgSend(this.id, OS.sel_appendBezierPathWithRoundedRect_xRadius_yRadius_, rect, xRadius, yRadius);
 }
 
 public static NSBezierPath bezierPath() {
-	int result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPath);
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPath);
 	return result != 0 ? new NSBezierPath(result) : null;
 }
 
 public NSBezierPath bezierPathByFlatteningPath() {
-	int result = OS.objc_msgSend(this.id, OS.sel_bezierPathByFlatteningPath);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_bezierPathByFlatteningPath);
 	return result == this.id ? this : (result != 0 ? new NSBezierPath(result) : null);
 }
 
 public static NSBezierPath bezierPathWithRect(NSRect rect) {
-	int result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRect_, rect);
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_bezierPathWithRect_, rect);
 	return result != 0 ? new NSBezierPath(result) : null;
 }
 
@@ -101,16 +101,16 @@ public void curveToPoint(NSPoint endPoint, NSPoint controlPoint1, NSPoint contro
 	OS.objc_msgSend(this.id, OS.sel_curveToPoint_controlPoint1_controlPoint2_, endPoint, controlPoint1, controlPoint2);
 }
 
-public static float defaultFlatness() {
+public static float /*double*/ defaultFlatness() {
 	return (float)OS.objc_msgSend_fpret(OS.class_NSBezierPath, OS.sel_defaultFlatness);
 }
 
-public int elementAtIndex(int index, int points) {
-	return OS.objc_msgSend(this.id, OS.sel_elementAtIndex_associatedPoints_, index, points);
+public int /*long*/ elementAtIndex(int /*long*/ index, int /*long*/ points) {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_elementAtIndex_associatedPoints_, index, points);
 }
 
-public int elementCount() {
-	return OS.objc_msgSend(this.id, OS.sel_elementCount);
+public int /*long*/ elementCount() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_elementCount);
 }
 
 public void fill() {
@@ -141,31 +141,31 @@ public void setClip() {
 	OS.objc_msgSend(this.id, OS.sel_setClip);
 }
 
-public static void setDefaultFlatness(float flatness) {
+public static void setDefaultFlatness(float /*double*/ flatness) {
 	OS.objc_msgSend(OS.class_NSBezierPath, OS.sel_setDefaultFlatness_, flatness);
 }
 
-public void setLineCapStyle(int lineCapStyle) {
+public void setLineCapStyle(int /*long*/ lineCapStyle) {
 	OS.objc_msgSend(this.id, OS.sel_setLineCapStyle_, lineCapStyle);
 }
 
-public void setLineDash(float[] pattern, int count, float phase) {
+public void setLineDash(float[] /*double[]*/ pattern, int /*long*/ count, float /*double*/ phase) {
 	OS.objc_msgSend(this.id, OS.sel_setLineDash_count_phase_, pattern, count, phase);
 }
 
-public void setLineJoinStyle(int lineJoinStyle) {
+public void setLineJoinStyle(int /*long*/ lineJoinStyle) {
 	OS.objc_msgSend(this.id, OS.sel_setLineJoinStyle_, lineJoinStyle);
 }
 
-public void setLineWidth(float lineWidth) {
+public void setLineWidth(float /*double*/ lineWidth) {
 	OS.objc_msgSend(this.id, OS.sel_setLineWidth_, lineWidth);
 }
 
-public void setMiterLimit(float miterLimit) {
+public void setMiterLimit(float /*double*/ miterLimit) {
 	OS.objc_msgSend(this.id, OS.sel_setMiterLimit_, miterLimit);
 }
 
-public void setWindingRule(int windingRule) {
+public void setWindingRule(int /*long*/ windingRule) {
 	OS.objc_msgSend(this.id, OS.sel_setWindingRule_, windingRule);
 }
